Changed the inherited method in Pawn to valid_move with a super call.
Per Sam's request, I have changed the inherited method in Pawn to
valid_move? and added a super call in pawn.rb. The logic for a pawn's
valid move is still in legal_move? in pawn.rb. I have also refactored
the move_to! method. It is now a little cleaner and easier to follow.